backend/requirements.txt
fastapi
uvicorn[standard]
watchdog
opencv-contrib-python
colour-science
numpy
Pillow
scipy
cv2-enumerate-cameras
